loadbalancer/maps: Relocate types from pkg/maps/lbmap
The pkg/maps/lbmap is being replaced by pkg/loadbalancer/maps.

Move the load-balancer BPF map Go types from pkg/maps/lbmap into
pkg/loadbalancer/maps. Add type aliases to pkg/maps/lbmap.
