+++ pkg/loadbalancer/maps/lbmaps.go
@@ -22,7 +22,6 @@ import (
 	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maglev"
-	"github.com/cilium/cilium/pkg/maps/lbmap"
 )
 
 type lbmapsParams struct {
@@ -67,39 +66,39 @@ func newLBMaps(p lbmapsParams) bpf.MapOut[LBMaps] {
 }
 
 type serviceMaps interface {
-	UpdateService(key lbmap.ServiceKey, value lbmap.ServiceValue) error
-	DeleteService(key lbmap.ServiceKey) error
-	DumpService(cb func(lbmap.ServiceKey, lbmap.ServiceValue)) error
+	UpdateService(key ServiceKey, value ServiceValue) error
+	DeleteService(key ServiceKey) error
+	DumpService(cb func(ServiceKey, ServiceValue)) error
 }
 
 type backendMaps interface {
-	UpdateBackend(lbmap.BackendKey, lbmap.BackendValue) error
-	DeleteBackend(lbmap.BackendKey) error
-	DumpBackend(cb func(lbmap.BackendKey, lbmap.BackendValue)) error
+	UpdateBackend(BackendKey, BackendValue) error
+	DeleteBackend(BackendKey) error
+	DumpBackend(cb func(BackendKey, BackendValue)) error
 }
 
 type revNatMaps interface {
-	UpdateRevNat(lbmap.RevNatKey, lbmap.RevNatValue) error
-	DeleteRevNat(lbmap.RevNatKey) error
-	DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) error
+	UpdateRevNat(RevNatKey, RevNatValue) error
+	DeleteRevNat(RevNatKey) error
+	DumpRevNat(cb func(RevNatKey, RevNatValue)) error
 }
 
 type affinityMaps interface {
-	UpdateAffinityMatch(*lbmap.AffinityMatchKey, *lbmap.AffinityMatchValue) error
-	DeleteAffinityMatch(*lbmap.AffinityMatchKey) error
-	DumpAffinityMatch(cb func(*lbmap.AffinityMatchKey, *lbmap.AffinityMatchValue)) error
+	UpdateAffinityMatch(*AffinityMatchKey, *AffinityMatchValue) error
+	DeleteAffinityMatch(*AffinityMatchKey) error
+	DumpAffinityMatch(cb func(*AffinityMatchKey, *AffinityMatchValue)) error
 }
 
 type sourceRangeMaps interface {
-	UpdateSourceRange(lbmap.SourceRangeKey, *lbmap.SourceRangeValue) error
-	DeleteSourceRange(lbmap.SourceRangeKey) error
-	DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.SourceRangeValue)) error
+	UpdateSourceRange(SourceRangeKey, *SourceRangeValue) error
+	DeleteSourceRange(SourceRangeKey) error
+	DumpSourceRange(cb func(SourceRangeKey, *SourceRangeValue)) error
 }
 
 type maglevMaps interface {
-	UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error
-	DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error
-	DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterVal, lbmap.MaglevInnerKey, *lbmap.MaglevInnerVal, bool)) error
+	UpdateMaglev(key MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error
+	DeleteMaglev(key MaglevOuterKey, ipv6 bool) error
+	DumpMaglev(cb func(MaglevOuterKey, MaglevOuterVal, MaglevInnerKey, *MaglevInnerVal, bool)) error
 }
 
 type sockRevNatMaps interface {
@@ -152,10 +151,10 @@ type BPFLBMaps struct {
 
 func newService4Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.Service4MapV2Name,
+		Service4MapV2Name,
 		ebpf.Hash,
-		&lbmap.Service4Key{},
-		&lbmap.Service4Value{},
+		&Service4Key{},
+		&Service4Value{},
 		maxEntries,
 		0,
 	)
@@ -163,10 +162,10 @@ func newService4Map(maxEntries int) *bpf.Map {
 
 func newService6Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.Service6MapV2Name,
+		Service6MapV2Name,
 		ebpf.Hash,
-		&lbmap.Service6Key{},
-		&lbmap.Service6Value{},
+		&Service6Key{},
+		&Service6Value{},
 		maxEntries,
 		0,
 	)
@@ -174,10 +173,10 @@ func newService6Map(maxEntries int) *bpf.Map {
 
 func newBackend4Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.Backend4MapV3Name,
+		Backend4MapV3Name,
 		ebpf.Hash,
-		&lbmap.Backend4KeyV3{},
-		&lbmap.Backend4ValueV3{},
+		&Backend4KeyV3{},
+		&Backend4ValueV3{},
 		maxEntries,
 		0,
 	)
@@ -185,10 +184,10 @@ func newBackend4Map(maxEntries int) *bpf.Map {
 
 func newBackend6Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.Backend6MapV3Name,
+		Backend6MapV3Name,
 		ebpf.Hash,
-		&lbmap.Backend6KeyV3{},
-		&lbmap.Backend6ValueV3{},
+		&Backend6KeyV3{},
+		&Backend6ValueV3{},
 		maxEntries,
 		0,
 	)
@@ -196,10 +195,10 @@ func newBackend6Map(maxEntries int) *bpf.Map {
 
 func newRevNat4Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.RevNat4MapName,
+		RevNat4MapName,
 		ebpf.Hash,
-		&lbmap.RevNat4Key{},
-		&lbmap.RevNat4Value{},
+		&RevNat4Key{},
+		&RevNat4Value{},
 		maxEntries,
 		0,
 	)
@@ -207,10 +206,10 @@ func newRevNat4Map(maxEntries int) *bpf.Map {
 
 func newRevNat6Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.RevNat6MapName,
+		RevNat6MapName,
 		ebpf.Hash,
-		&lbmap.RevNat6Key{},
-		&lbmap.RevNat6Value{},
+		&RevNat6Key{},
+		&RevNat6Value{},
 		maxEntries,
 		0,
 	)
@@ -218,10 +217,10 @@ func newRevNat6Map(maxEntries int) *bpf.Map {
 
 func newAffinityMatchMap(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.AffinityMatchMapName,
+		AffinityMatchMapName,
 		ebpf.Hash,
-		&lbmap.AffinityMatchKey{},
-		&lbmap.AffinityMatchValue{},
+		&AffinityMatchKey{},
+		&AffinityMatchValue{},
 		maxEntries,
 		0,
 	)
@@ -229,10 +228,10 @@ func newAffinityMatchMap(maxEntries int) *bpf.Map {
 
 func newSourceRange4Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.SourceRange4MapName,
+		SourceRange4MapName,
 		ebpf.LPMTrie,
-		&lbmap.SourceRangeKey4{},
-		&lbmap.SourceRangeValue{},
+		&SourceRangeKey4{},
+		&SourceRangeValue{},
 		maxEntries,
 		0,
 	)
@@ -240,10 +239,10 @@ func newSourceRange4Map(maxEntries int) *bpf.Map {
 
 func newSourceRange6Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.SourceRange6MapName,
+		SourceRange6MapName,
 		ebpf.LPMTrie,
-		&lbmap.SourceRangeKey6{},
-		&lbmap.SourceRangeValue{},
+		&SourceRangeKey6{},
+		&SourceRangeValue{},
 		maxEntries,
 		0,
 	)
@@ -251,10 +250,10 @@ func newSourceRange6Map(maxEntries int) *bpf.Map {
 
 func newSockRevNat4Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.SockRevNat4MapName,
+		SockRevNat4MapName,
 		ebpf.LRUHash,
-		&lbmap.SockRevNat4Key{},
-		&lbmap.SockRevNat4Value{},
+		&SockRevNat4Key{},
+		&SockRevNat4Value{},
 		maxEntries,
 		0,
 	)
@@ -262,10 +261,10 @@ func newSockRevNat4Map(maxEntries int) *bpf.Map {
 
 func newSockRevNat6Map(maxEntries int) *bpf.Map {
 	return bpf.NewMap(
-		lbmap.SockRevNat6MapName,
+		SockRevNat6MapName,
 		ebpf.LRUHash,
-		&lbmap.SockRevNat6Key{},
-		&lbmap.SockRevNat6Value{},
+		&SockRevNat6Key{},
+		&SockRevNat6Value{},
 		maxEntries,
 		0,
 	)
@@ -275,8 +274,8 @@ func newMaglevOuterMap(name string, maxEntries int, innerSpec *ebpf.MapSpec) *bp
 	return bpf.NewMapWithInnerSpec(
 		name,
 		ebpf.HashOfMaps,
-		&lbmap.MaglevOuterKey{},
-		&lbmap.MaglevOuterVal{},
+		&MaglevOuterKey{},
+		&MaglevOuterVal{},
 		maxEntries,
 		0,
 		innerSpec.Copy(),
@@ -291,10 +290,10 @@ type mapDesc struct {
 
 func (r *BPFLBMaps) allMaps() ([]mapDesc, []mapDesc) {
 	newMaglev4 := func(maxEntries int) *bpf.Map {
-		return newMaglevOuterMap(lbmap.MaglevOuter4MapName, maxEntries, r.maglevInnerMapSpec)
+		return newMaglevOuterMap(MaglevOuter4MapName, maxEntries, r.maglevInnerMapSpec)
 	}
 	newMaglev6 := func(maxEntries int) *bpf.Map {
-		return newMaglevOuterMap(lbmap.MaglevOuter6MapName, maxEntries, r.maglevInnerMapSpec)
+		return newMaglevOuterMap(MaglevOuter6MapName, maxEntries, r.maglevInnerMapSpec)
 	}
 	v4Maps := []mapDesc{
 		{&r.service4Map, newService4Map, r.Cfg.LBServiceMapEntries},
@@ -332,11 +331,11 @@ func (r *BPFLBMaps) allMaps() ([]mapDesc, []mapDesc) {
 // Start implements cell.HookInterface.
 func (r *BPFLBMaps) Start(ctx cell.HookContext) (err error) {
 	r.maglevInnerMapSpec = &ebpf.MapSpec{
-		Name:       lbmap.MaglevInnerMapName,
+		Name:       MaglevInnerMapName,
 		Type:       ebpf.Array,
-		KeySize:    uint32(unsafe.Sizeof(lbmap.MaglevInnerKey{})),
+		KeySize:    uint32(unsafe.Sizeof(MaglevInnerKey{})),
 		MaxEntries: 1,
-		ValueSize:  lbmap.MaglevBackendLen * uint32(r.MaglevCfg.TableSize),
+		ValueSize:  MaglevBackendLen * uint32(r.MaglevCfg.TableSize),
 	}
 
 	mapsToCreate, mapsToDelete := r.allMaps()
@@ -413,12 +412,12 @@ func dumpMap[K bpf.MapKey, V bpf.MapValue](m *bpf.Map, cb func(K, V)) error {
 }
 
 // DeleteRevNat implements lbmaps.
-func (r *BPFLBMaps) DeleteRevNat(key lbmap.RevNatKey) error {
+func (r *BPFLBMaps) DeleteRevNat(key RevNatKey) error {
 	var err error
 	switch key.(type) {
-	case *lbmap.RevNat4Key:
+	case *RevNat4Key:
 		err = r.revNat4Map.Delete(key)
-	case *lbmap.RevNat6Key:
+	case *RevNat6Key:
 		err = r.revNat6Map.Delete(key)
 	default:
 		panic("unknown RevNatKey")
@@ -430,7 +429,7 @@ func (r *BPFLBMaps) DeleteRevNat(key lbmap.RevNatKey) error {
 }
 
 // DumpRevNat implements lbmaps.
-func (r *BPFLBMaps) DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) error {
+func (r *BPFLBMaps) DumpRevNat(cb func(RevNatKey, RevNatValue)) error {
 	return errors.Join(
 		dumpMap(r.revNat4Map, cb),
 		dumpMap(r.revNat6Map, cb),
@@ -438,11 +437,11 @@ func (r *BPFLBMaps) DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) erro
 }
 
 // UpdateRevNat4 implements lbmaps.
-func (r *BPFLBMaps) UpdateRevNat(key lbmap.RevNatKey, value lbmap.RevNatValue) error {
+func (r *BPFLBMaps) UpdateRevNat(key RevNatKey, value RevNatValue) error {
 	switch key.(type) {
-	case *lbmap.RevNat4Key:
+	case *RevNat4Key:
 		return r.revNat4Map.Update(key, value)
-	case *lbmap.RevNat6Key:
+	case *RevNat6Key:
 		return r.revNat6Map.Update(key, value)
 	default:
 		panic("unknown RevNatKey")
@@ -450,7 +449,7 @@ func (r *BPFLBMaps) UpdateRevNat(key lbmap.RevNatKey, value lbmap.RevNatValue) e
 }
 
 // DumpBackend implements lbmaps.
-func (r *BPFLBMaps) DumpBackend(cb func(lbmap.BackendKey, lbmap.BackendValue)) error {
+func (r *BPFLBMaps) DumpBackend(cb func(BackendKey, BackendValue)) error {
 	return errors.Join(
 		dumpMap(r.backend4Map, cb),
 		dumpMap(r.backend6Map, cb),
@@ -458,12 +457,12 @@ func (r *BPFLBMaps) DumpBackend(cb func(lbmap.BackendKey, lbmap.BackendValue)) e
 }
 
 // DeleteBackend implements lbmaps.
-func (r *BPFLBMaps) DeleteBackend(key lbmap.BackendKey) error {
+func (r *BPFLBMaps) DeleteBackend(key BackendKey) error {
 	var err error
 	switch key.(type) {
-	case *lbmap.Backend4KeyV3:
+	case *Backend4KeyV3:
 		err = r.backend4Map.Delete(key)
-	case *lbmap.Backend6KeyV3:
+	case *Backend6KeyV3:
 		err = r.backend6Map.Delete(key)
 	default:
 		panic("unknown BackendKey")
@@ -475,12 +474,12 @@ func (r *BPFLBMaps) DeleteBackend(key lbmap.BackendKey) error {
 }
 
 // DeleteService implements lbmaps.
-func (r *BPFLBMaps) DeleteService(key lbmap.ServiceKey) error {
+func (r *BPFLBMaps) DeleteService(key ServiceKey) error {
 	var err error
 	switch key.(type) {
-	case *lbmap.Service4Key:
+	case *Service4Key:
 		_, err = r.service4Map.SilentDelete(key)
-	case *lbmap.Service6Key:
+	case *Service6Key:
 		_, err = r.service6Map.SilentDelete(key)
 	default:
 		panic("unknown ServiceKey")
@@ -489,7 +488,7 @@ func (r *BPFLBMaps) DeleteService(key lbmap.ServiceKey) error {
 }
 
 // DumpService implements lbmaps.
-func (r *BPFLBMaps) DumpService(cb func(lbmap.ServiceKey, lbmap.ServiceValue)) error {
+func (r *BPFLBMaps) DumpService(cb func(ServiceKey, ServiceValue)) error {
 	return errors.Join(
 		dumpMap(r.service4Map, cb),
 		dumpMap(r.service6Map, cb),
@@ -497,11 +496,11 @@ func (r *BPFLBMaps) DumpService(cb func(lbmap.ServiceKey, lbmap.ServiceValue)) e
 }
 
 // UpdateBackend implements lbmaps.
-func (r *BPFLBMaps) UpdateBackend(key lbmap.BackendKey, value lbmap.BackendValue) error {
+func (r *BPFLBMaps) UpdateBackend(key BackendKey, value BackendValue) error {
 	switch key.(type) {
-	case *lbmap.Backend4KeyV3:
+	case *Backend4KeyV3:
 		return r.backend4Map.Update(key, value)
-	case *lbmap.Backend6KeyV3:
+	case *Backend6KeyV3:
 		return r.backend6Map.Update(key, value)
 	default:
 		panic("unknown BackendKey")
@@ -509,11 +508,11 @@ func (r *BPFLBMaps) UpdateBackend(key lbmap.BackendKey, value lbmap.BackendValue
 }
 
 // UpdateService implements lbmaps.
-func (r *BPFLBMaps) UpdateService(key lbmap.ServiceKey, value lbmap.ServiceValue) error {
+func (r *BPFLBMaps) UpdateService(key ServiceKey, value ServiceValue) error {
 	switch key.(type) {
-	case *lbmap.Service4Key:
+	case *Service4Key:
 		return r.service4Map.Update(key, value)
-	case *lbmap.Service6Key:
+	case *Service6Key:
 		return r.service6Map.Update(key, value)
 	default:
 		panic("unknown ServiceKey")
@@ -521,7 +520,7 @@ func (r *BPFLBMaps) UpdateService(key lbmap.ServiceKey, value lbmap.ServiceValue
 }
 
 // DeleteAffinityMatch implements lbmaps.
-func (r *BPFLBMaps) DeleteAffinityMatch(key *lbmap.AffinityMatchKey) error {
+func (r *BPFLBMaps) DeleteAffinityMatch(key *AffinityMatchKey) error {
 	err := r.affinityMatchMap.Delete(key)
 	if errors.Is(err, ebpf.ErrKeyNotExist) {
 		return nil
@@ -530,22 +529,22 @@ func (r *BPFLBMaps) DeleteAffinityMatch(key *lbmap.AffinityMatchKey) error {
 }
 
 // DumpAffinityMatch implements lbmaps.
-func (r *BPFLBMaps) DumpAffinityMatch(cb func(*lbmap.AffinityMatchKey, *lbmap.AffinityMatchValue)) error {
+func (r *BPFLBMaps) DumpAffinityMatch(cb func(*AffinityMatchKey, *AffinityMatchValue)) error {
 	return dumpMap(r.affinityMatchMap, cb)
 }
 
 // UpdateAffinityMatch implements lbmaps.
-func (r *BPFLBMaps) UpdateAffinityMatch(key *lbmap.AffinityMatchKey, value *lbmap.AffinityMatchValue) error {
+func (r *BPFLBMaps) UpdateAffinityMatch(key *AffinityMatchKey, value *AffinityMatchValue) error {
 	return r.affinityMatchMap.Update(key, value)
 }
 
 // DeleteSourceRange implements lbmaps.
-func (r *BPFLBMaps) DeleteSourceRange(key lbmap.SourceRangeKey) error {
+func (r *BPFLBMaps) DeleteSourceRange(key SourceRangeKey) error {
 	var err error
 	switch key.(type) {
-	case *lbmap.SourceRangeKey4:
+	case *SourceRangeKey4:
 		err = r.sourceRange4Map.Delete(key)
-	case *lbmap.SourceRangeKey6:
+	case *SourceRangeKey6:
 		err = r.sourceRange6Map.Delete(key)
 	default:
 		panic("unknown SourceRangeKey")
@@ -557,7 +556,7 @@ func (r *BPFLBMaps) DeleteSourceRange(key lbmap.SourceRangeKey) error {
 }
 
 // DumpSourceRange implements lbmaps.
-func (r *BPFLBMaps) DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.SourceRangeValue)) error {
+func (r *BPFLBMaps) DumpSourceRange(cb func(SourceRangeKey, *SourceRangeValue)) error {
 	return errors.Join(
 		dumpMap(r.sourceRange4Map, cb),
 		dumpMap(r.sourceRange6Map, cb),
@@ -565,11 +564,11 @@ func (r *BPFLBMaps) DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.SourceR
 }
 
 // UpdateSourceRange implements lbmaps.
-func (r *BPFLBMaps) UpdateSourceRange(key lbmap.SourceRangeKey, value *lbmap.SourceRangeValue) error {
+func (r *BPFLBMaps) UpdateSourceRange(key SourceRangeKey, value *SourceRangeValue) error {
 	switch key.(type) {
-	case *lbmap.SourceRangeKey4:
+	case *SourceRangeKey4:
 		return r.sourceRange4Map.Update(key, value)
-	case *lbmap.SourceRangeKey6:
+	case *SourceRangeKey6:
 		return r.sourceRange6Map.Update(key, value)
 	default:
 		panic("unknown SourceRangeKey")
@@ -577,20 +576,20 @@ func (r *BPFLBMaps) UpdateSourceRange(key lbmap.SourceRangeKey, value *lbmap.Sou
 }
 
 // UpdateMaglev implements lbmaps.
-func (r *BPFLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
+func (r *BPFLBMaps) UpdateMaglev(key MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
 	inner, err := ebpf.NewMap(r.maglevInnerMapSpec)
 	if err != nil {
 		return fmt.Errorf("failed to create map %q: %w", r.maglevInnerMapSpec.Name, err)
 	}
 	defer inner.Close()
-	var singletonKey lbmap.MaglevInnerKey
+	var singletonKey MaglevInnerKey
 	if err := inner.Update(singletonKey, backendIDs, 0); err != nil {
 		return fmt.Errorf("updating backends: %w", err)
 	}
-	outerKey := &lbmap.MaglevOuterKey{
+	outerKey := &MaglevOuterKey{
 		RevNatID: byteorder.HostToNetwork16(key.RevNatID),
 	}
-	outerValue := &lbmap.MaglevOuterVal{FD: uint32(inner.FD())}
+	outerValue := &MaglevOuterVal{FD: uint32(inner.FD())}
 	if ipv6 {
 		return r.maglev6Map.Update(outerKey, outerValue)
 	} else {
@@ -599,8 +598,8 @@ func (r *BPFLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbala
 }
 
 // DeleteMaglev implements lbmaps.
-func (r *BPFLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
-	outerKey := &lbmap.MaglevOuterKey{
+func (r *BPFLBMaps) DeleteMaglev(key MaglevOuterKey, ipv6 bool) error {
+	outerKey := &MaglevOuterKey{
 		RevNatID: byteorder.HostToNetwork16(key.RevNatID),
 	}
 	ebpfmap := r.maglev4Map
@@ -614,22 +613,22 @@ func (r *BPFLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
 	return err
 }
 
-func (r *BPFLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterVal, lbmap.MaglevInnerKey, *lbmap.MaglevInnerVal, bool)) error {
+func (r *BPFLBMaps) DumpMaglev(cb func(MaglevOuterKey, MaglevOuterVal, MaglevInnerKey, *MaglevInnerVal, bool)) error {
 	var errs []error
 	var ipv6 bool
 	cbWrap := func(key bpf.MapKey, value bpf.MapValue) {
-		maglevKey := lbmap.MaglevOuterKey{
-			RevNatID: byteorder.NetworkToHost16(key.(*lbmap.MaglevOuterKey).RevNatID),
+		maglevKey := MaglevOuterKey{
+			RevNatID: byteorder.NetworkToHost16(key.(*MaglevOuterKey).RevNatID),
 		}
-		maglevValue := value.(*lbmap.MaglevOuterVal)
-		inner, err := lbmap.MaglevInnerMapFromID(r.Log, maglevValue.FD)
+		maglevValue := value.(*MaglevOuterVal)
+		inner, err := MaglevInnerMapFromID(r.Log, maglevValue.FD)
 		if err != nil {
 			errs = append(errs, fmt.Errorf("cannot open inner map with fd %d: %w", maglevValue.FD, err))
 			return
 		}
 		defer inner.Close()
 		// Maglev inner map has a single key and a huge value.
-		var singletonKey lbmap.MaglevInnerKey
+		var singletonKey MaglevInnerKey
 		innerValue, err := inner.Lookup(&singletonKey)
 		if err != nil {
 			errs = append(errs, fmt.Errorf("cannot look up backends in inner map with id %d: %w", maglevValue.FD, err))
@@ -650,11 +649,11 @@ func (r *BPFLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterVa
 // DeleteSockRevNat implements LBMaps.
 func (r *BPFLBMaps) DeleteSockRevNat(cookie uint64, addr net.IP, port uint16) error {
 	if addr.To4() != nil && r.sockRevNat4Map != nil {
-		key := lbmap.NewSockRevNat4Key(cookie, addr, port)
+		key := NewSockRevNat4Key(cookie, addr, port)
 		_, err := r.sockRevNat4Map.SilentDelete(key)
 		return err
 	} else if r.sockRevNat6Map != nil {
-		key := lbmap.NewSockRevNat6Key(cookie, addr, port)
+		key := NewSockRevNat6Key(cookie, addr, port)
 		_, err := r.sockRevNat6Map.SilentDelete(key)
 		return err
 	}
@@ -664,16 +663,16 @@ func (r *BPFLBMaps) DeleteSockRevNat(cookie uint64, addr net.IP, port uint16) er
 // UpdateSockRevNat implements LBMaps.
 func (r *BPFLBMaps) UpdateSockRevNat(cookie uint64, addr net.IP, port uint16, revNatIndex uint16) error {
 	if addr.To4() != nil && r.sockRevNat4Map != nil {
-		key := lbmap.NewSockRevNat4Key(cookie, addr, port)
-		value := lbmap.SockRevNat4Value{
+		key := NewSockRevNat4Key(cookie, addr, port)
+		value := SockRevNat4Value{
 			Address:     key.Address,
 			Port:        key.Port,
 			RevNatIndex: revNatIndex,
 		}
 		return r.sockRevNat4Map.Update(key, &value)
 	} else if r.sockRevNat6Map != nil {
-		key := lbmap.NewSockRevNat6Key(cookie, addr, port)
-		value := lbmap.SockRevNat6Value{
+		key := NewSockRevNat6Key(cookie, addr, port)
+		value := SockRevNat6Value{
 			Address:     key.Address,
 			Port:        key.Port,
 			RevNatIndex: revNatIndex,
@@ -685,17 +684,66 @@ func (r *BPFLBMaps) UpdateSockRevNat(cookie uint64, addr net.IP, port uint16, re
 
 func (r *BPFLBMaps) ExistsSockRevNat(cookie uint64, addr net.IP, port uint16) bool {
 	if addr.To4() != nil && r.sockRevNat4Map != nil {
-		key := lbmap.NewSockRevNat4Key(cookie, addr, port)
+		key := NewSockRevNat4Key(cookie, addr, port)
 		if v, _ := r.sockRevNat4Map.Lookup(key); v != nil {
 			return true
 		}
 	} else if r.sockRevNat6Map != nil {
-		key := lbmap.NewSockRevNat6Key(cookie, addr, port)
+		key := NewSockRevNat6Key(cookie, addr, port)
 		if v, _ := r.sockRevNat6Map.Lookup(key); v != nil {
 			return true
 		}
 	}
 	return false
+}
+
+// MaglevInnerMap represents a maglev inner map.
+type MaglevInnerMap struct {
+	*ebpf.Map
+}
+
+// TableSize returns the amount of backends this map can hold as a value.
+func (m MaglevInnerMap) TableSize() uint32 {
+	return m.Map.ValueSize() / uint32(MaglevBackendLen)
+}
+
+// UpdateBackends updates the maglev inner map's list of backends.
+func (m MaglevInnerMap) UpdateBackends(backends []loadbalancer.BackendID) error {
+	// Backends are stored at inner map key zero.
+	var key MaglevInnerKey
+	return m.Map.Update(key, backends, 0)
+}
+
+// MaglevInnerMapFromID returns a new object representing the maglev inner map
+// identified by an ID.
+func MaglevInnerMapFromID(log *slog.Logger, id uint32) (MaglevInnerMap, error) {
+	m, err := ebpf.NewMapFromID(ebpf.MapID(id))
+	return MaglevInnerMap{m}, err
+}
+
+// Lookup returns the value associated with a given key for a maglev inner map.
+func (m MaglevInnerMap) Lookup(key *MaglevInnerKey) (*MaglevInnerVal, error) {
+	value := &MaglevInnerVal{
+		BackendIDs: make([]loadbalancer.BackendID, m.TableSize()),
+	}
+
+	if err := m.Map.Lookup(key, &value.BackendIDs); err != nil {
+		return nil, err
+	}
+
+	return value, nil
+}
+
+// DumpBackends returns the first key of the map as stringified ints for dumping purposes.
+func (m MaglevInnerMap) DumpBackends() (string, error) {
+	// A service's backend array sits at the first key of the inner map.
+	var key MaglevInnerKey
+	val, err := m.Lookup(&key)
+	if err != nil {
+		return "", fmt.Errorf("lookup up first inner map key (backends): %w", err)
+	}
+
+	return fmt.Sprintf("%v", val.BackendIDs), nil
 }
 
 // IsEmpty implements lbmaps.
@@ -742,7 +790,7 @@ func (f *FaultyLBMaps) UpdateSockRevNat(cookie uint64, addr net.IP, port uint16,
 }
 
 // DeleteSourceRange implements lbmaps.
-func (f *FaultyLBMaps) DeleteSourceRange(key lbmap.SourceRangeKey) error {
+func (f *FaultyLBMaps) DeleteSourceRange(key SourceRangeKey) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -750,7 +798,7 @@ func (f *FaultyLBMaps) DeleteSourceRange(key lbmap.SourceRangeKey) error {
 }
 
 // DumpSourceRange implements lbmaps.
-func (f *FaultyLBMaps) DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.SourceRangeValue)) error {
+func (f *FaultyLBMaps) DumpSourceRange(cb func(SourceRangeKey, *SourceRangeValue)) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -758,7 +806,7 @@ func (f *FaultyLBMaps) DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.Sour
 }
 
 // UpdateSourceRange implements lbmaps.
-func (f *FaultyLBMaps) UpdateSourceRange(key lbmap.SourceRangeKey, value *lbmap.SourceRangeValue) error {
+func (f *FaultyLBMaps) UpdateSourceRange(key SourceRangeKey, value *SourceRangeValue) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -766,7 +814,7 @@ func (f *FaultyLBMaps) UpdateSourceRange(key lbmap.SourceRangeKey, value *lbmap.
 }
 
 // DeleteAffinityMatch implements lbmaps.
-func (f *FaultyLBMaps) DeleteAffinityMatch(key *lbmap.AffinityMatchKey) error {
+func (f *FaultyLBMaps) DeleteAffinityMatch(key *AffinityMatchKey) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -774,12 +822,12 @@ func (f *FaultyLBMaps) DeleteAffinityMatch(key *lbmap.AffinityMatchKey) error {
 }
 
 // DumpAffinityMatch implements lbmaps.
-func (f *FaultyLBMaps) DumpAffinityMatch(cb func(*lbmap.AffinityMatchKey, *lbmap.AffinityMatchValue)) error {
+func (f *FaultyLBMaps) DumpAffinityMatch(cb func(*AffinityMatchKey, *AffinityMatchValue)) error {
 	return f.impl.DumpAffinityMatch(cb)
 }
 
 // UpdateAffinityMatch implements lbmaps.
-func (f *FaultyLBMaps) UpdateAffinityMatch(key *lbmap.AffinityMatchKey, value *lbmap.AffinityMatchValue) error {
+func (f *FaultyLBMaps) UpdateAffinityMatch(key *AffinityMatchKey, value *AffinityMatchValue) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -787,7 +835,7 @@ func (f *FaultyLBMaps) UpdateAffinityMatch(key *lbmap.AffinityMatchKey, value *l
 }
 
 // DeleteRevNat implements lbmaps.
-func (f *FaultyLBMaps) DeleteRevNat(key lbmap.RevNatKey) error {
+func (f *FaultyLBMaps) DeleteRevNat(key RevNatKey) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -795,7 +843,7 @@ func (f *FaultyLBMaps) DeleteRevNat(key lbmap.RevNatKey) error {
 }
 
 // DumpRevNat implements lbmaps.
-func (f *FaultyLBMaps) DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) error {
+func (f *FaultyLBMaps) DumpRevNat(cb func(RevNatKey, RevNatValue)) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -803,7 +851,7 @@ func (f *FaultyLBMaps) DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) e
 }
 
 // UpdateRevNat implements lbmaps.
-func (f *FaultyLBMaps) UpdateRevNat(key lbmap.RevNatKey, value lbmap.RevNatValue) error {
+func (f *FaultyLBMaps) UpdateRevNat(key RevNatKey, value RevNatValue) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -811,7 +859,7 @@ func (f *FaultyLBMaps) UpdateRevNat(key lbmap.RevNatKey, value lbmap.RevNatValue
 }
 
 // DeleteBackend implements lbmaps.
-func (f *FaultyLBMaps) DeleteBackend(key lbmap.BackendKey) error {
+func (f *FaultyLBMaps) DeleteBackend(key BackendKey) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -819,7 +867,7 @@ func (f *FaultyLBMaps) DeleteBackend(key lbmap.BackendKey) error {
 }
 
 // DeleteService implements lbmaps.
-func (f *FaultyLBMaps) DeleteService(key lbmap.ServiceKey) error {
+func (f *FaultyLBMaps) DeleteService(key ServiceKey) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -827,12 +875,12 @@ func (f *FaultyLBMaps) DeleteService(key lbmap.ServiceKey) error {
 }
 
 // DumpBackend implements lbmaps.
-func (f *FaultyLBMaps) DumpBackend(cb func(lbmap.BackendKey, lbmap.BackendValue)) error {
+func (f *FaultyLBMaps) DumpBackend(cb func(BackendKey, BackendValue)) error {
 	return f.impl.DumpBackend(cb)
 }
 
 // DumpService implements lbmaps.
-func (f *FaultyLBMaps) DumpService(cb func(lbmap.ServiceKey, lbmap.ServiceValue)) error {
+func (f *FaultyLBMaps) DumpService(cb func(ServiceKey, ServiceValue)) error {
 	return f.impl.DumpService(cb)
 }
 
@@ -842,7 +890,7 @@ func (f *FaultyLBMaps) IsEmpty() bool {
 }
 
 // UpdateBackend implements lbmaps.
-func (f *FaultyLBMaps) UpdateBackend(key lbmap.BackendKey, value lbmap.BackendValue) error {
+func (f *FaultyLBMaps) UpdateBackend(key BackendKey, value BackendValue) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -850,7 +898,7 @@ func (f *FaultyLBMaps) UpdateBackend(key lbmap.BackendKey, value lbmap.BackendVa
 }
 
 // UpdateService implements lbmaps.
-func (f *FaultyLBMaps) UpdateService(key lbmap.ServiceKey, value lbmap.ServiceValue) error {
+func (f *FaultyLBMaps) UpdateService(key ServiceKey, value ServiceValue) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -858,7 +906,7 @@ func (f *FaultyLBMaps) UpdateService(key lbmap.ServiceKey, value lbmap.ServiceVa
 }
 
 // UpdateMaglev implements lbmaps.
-func (f *FaultyLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
+func (f *FaultyLBMaps) UpdateMaglev(key MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -866,7 +914,7 @@ func (f *FaultyLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadb
 }
 
 // DeleteMaglev implements lbmaps.
-func (f *FaultyLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
+func (f *FaultyLBMaps) DeleteMaglev(key MaglevOuterKey, ipv6 bool) error {
 	if f.isFaulty() {
 		return errFaulty
 	}
@@ -874,7 +922,7 @@ func (f *FaultyLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
 }
 
 // DumpMaglev implements lbmaps.
-func (f *FaultyLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterVal, lbmap.MaglevInnerKey, *lbmap.MaglevInnerVal, bool)) error {
+func (f *FaultyLBMaps) DumpMaglev(cb func(MaglevOuterKey, MaglevOuterVal, MaglevInnerKey, *MaglevInnerVal, bool)) error {
 	return f.impl.DumpMaglev(cb)
 }
 
@@ -950,99 +998,99 @@ func NewFakeLBMaps() LBMaps {
 }
 
 // DeleteAffinityMatch implements lbmaps.
-func (f *FakeLBMaps) DeleteAffinityMatch(key *lbmap.AffinityMatchKey) error {
+func (f *FakeLBMaps) DeleteAffinityMatch(key *AffinityMatchKey) error {
 	return f.aff.delete(key)
 }
 
 // DeleteBackend implements lbmaps.
-func (f *FakeLBMaps) DeleteBackend(key lbmap.BackendKey) error {
+func (f *FakeLBMaps) DeleteBackend(key BackendKey) error {
 	return f.be.delete(key)
 }
 
 // DeleteRevNat implements lbmaps.
-func (f *FakeLBMaps) DeleteRevNat(key lbmap.RevNatKey) error {
+func (f *FakeLBMaps) DeleteRevNat(key RevNatKey) error {
 	return f.revNat.delete(key)
 }
 
 // DeleteService implements lbmaps.
-func (f *FakeLBMaps) DeleteService(key lbmap.ServiceKey) error {
+func (f *FakeLBMaps) DeleteService(key ServiceKey) error {
 	return f.svc.delete(key)
 }
 
 // DeleteSourceRange implements lbmaps.
-func (f *FakeLBMaps) DeleteSourceRange(key lbmap.SourceRangeKey) error {
+func (f *FakeLBMaps) DeleteSourceRange(key SourceRangeKey) error {
 	return f.srcRange.delete(key)
 }
 
 // DumpAffinityMatch implements lbmaps.
-func (f *FakeLBMaps) DumpAffinityMatch(cb func(*lbmap.AffinityMatchKey, *lbmap.AffinityMatchValue)) error {
+func (f *FakeLBMaps) DumpAffinityMatch(cb func(*AffinityMatchKey, *AffinityMatchValue)) error {
 	dumpFakeBPFMap(&f.aff, cb)
 	return nil
 }
 
 // DumpBackend implements lbmaps.
-func (f *FakeLBMaps) DumpBackend(cb func(lbmap.BackendKey, lbmap.BackendValue)) error {
+func (f *FakeLBMaps) DumpBackend(cb func(BackendKey, BackendValue)) error {
 	dumpFakeBPFMap(&f.be, cb)
 	return nil
 }
 
 // DumpRevNat implements lbmaps.
-func (f *FakeLBMaps) DumpRevNat(cb func(lbmap.RevNatKey, lbmap.RevNatValue)) error {
+func (f *FakeLBMaps) DumpRevNat(cb func(RevNatKey, RevNatValue)) error {
 	dumpFakeBPFMap(&f.revNat, cb)
 	return nil
 }
 
 // DumpService implements lbmaps.
-func (f *FakeLBMaps) DumpService(cb func(lbmap.ServiceKey, lbmap.ServiceValue)) error {
+func (f *FakeLBMaps) DumpService(cb func(ServiceKey, ServiceValue)) error {
 	dumpFakeBPFMap(&f.svc, cb)
 	return nil
 }
 
 // DumpSourceRange implements lbmaps.
-func (f *FakeLBMaps) DumpSourceRange(cb func(lbmap.SourceRangeKey, *lbmap.SourceRangeValue)) error {
+func (f *FakeLBMaps) DumpSourceRange(cb func(SourceRangeKey, *SourceRangeValue)) error {
 	dumpFakeBPFMap(&f.srcRange, cb)
 	return nil
 }
 
 // UpdateAffinityMatch implements lbmaps.
-func (f *FakeLBMaps) UpdateAffinityMatch(key *lbmap.AffinityMatchKey, value *lbmap.AffinityMatchValue) error {
+func (f *FakeLBMaps) UpdateAffinityMatch(key *AffinityMatchKey, value *AffinityMatchValue) error {
 	return f.aff.update(key, value)
 }
 
 // UpdateBackend implements lbmaps.
-func (f *FakeLBMaps) UpdateBackend(key lbmap.BackendKey, value lbmap.BackendValue) error {
+func (f *FakeLBMaps) UpdateBackend(key BackendKey, value BackendValue) error {
 	return f.be.update(key, value)
 }
 
 // UpdateRevNat implements lbmaps.
-func (f *FakeLBMaps) UpdateRevNat(key lbmap.RevNatKey, value lbmap.RevNatValue) error {
+func (f *FakeLBMaps) UpdateRevNat(key RevNatKey, value RevNatValue) error {
 	return f.revNat.update(key, value)
 }
 
 // UpdateService implements lbmaps.
-func (f *FakeLBMaps) UpdateService(key lbmap.ServiceKey, value lbmap.ServiceValue) error {
+func (f *FakeLBMaps) UpdateService(key ServiceKey, value ServiceValue) error {
 	return f.svc.update(key, value)
 }
 
 // UpdateSourceRange implements lbmaps.
-func (f *FakeLBMaps) UpdateSourceRange(key lbmap.SourceRangeKey, value *lbmap.SourceRangeValue) error {
+func (f *FakeLBMaps) UpdateSourceRange(key SourceRangeKey, value *SourceRangeValue) error {
 	return f.srcRange.update(key, value)
 }
 
 // UpdateMaglev implements lbmaps.
-func (f *FakeLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
+func (f *FakeLBMaps) UpdateMaglev(key MaglevOuterKey, backendIDs []loadbalancer.BackendID, ipv6 bool) error {
 	var outer *fakeBPFMap
 	if ipv6 {
 		outer = &f.mglv6
 	} else {
 		outer = &f.mglv4
 	}
-	var singletonKey lbmap.MaglevInnerKey
+	var singletonKey MaglevInnerKey
 	inner := &fakeBPFMap{}
 	currentID := f.nextID
 	f.nextID++
 	f.inners.Store(currentID, inner)
-	value := lbmap.MaglevOuterVal{
+	value := MaglevOuterVal{
 		FD: currentID,
 	}
 	if err := inner.update(&singletonKey, backendIDs); err != nil {
@@ -1055,7 +1103,7 @@ func (f *FakeLBMaps) UpdateMaglev(key lbmap.MaglevOuterKey, backendIDs []loadbal
 }
 
 // DeleteMaglev implements lbmaps.
-func (f *FakeLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
+func (f *FakeLBMaps) DeleteMaglev(key MaglevOuterKey, ipv6 bool) error {
 	if ipv6 {
 		return f.mglv6.delete(&key)
 	} else {
@@ -1063,10 +1111,10 @@ func (f *FakeLBMaps) DeleteMaglev(key lbmap.MaglevOuterKey, ipv6 bool) error {
 	}
 }
 
-func (f *FakeLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterVal, lbmap.MaglevInnerKey, *lbmap.MaglevInnerVal, bool)) error {
+func (f *FakeLBMaps) DumpMaglev(cb func(MaglevOuterKey, MaglevOuterVal, MaglevInnerKey, *MaglevInnerVal, bool)) error {
 	var err error
-	cbWrap := func(key lbmap.MaglevOuterKey, value lbmap.MaglevOuterVal, ipv6 bool) bool {
-		singletonKey := lbmap.MaglevInnerKey{}
+	cbWrap := func(key MaglevOuterKey, value MaglevOuterVal, ipv6 bool) bool {
+		singletonKey := MaglevInnerKey{}
 		innerMap, ok := f.inners.Load(value.FD)
 		if !ok {
 			err = fmt.Errorf("inner map %d not found", value.FD)
@@ -1077,14 +1125,14 @@ func (f *FakeLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterV
 			err = fmt.Errorf("failed to fetch the value from the inner map for RevNatID=%d and FD=%d", key.RevNatID, value.FD)
 			return false
 		}
-		cb(key, value, *innerValue.a.(*lbmap.MaglevInnerKey), &lbmap.MaglevInnerVal{BackendIDs: innerValue.b.([]loadbalancer.BackendID)}, ipv6)
+		cb(key, value, *innerValue.a.(*MaglevInnerKey), &MaglevInnerVal{BackendIDs: innerValue.b.([]loadbalancer.BackendID)}, ipv6)
 		return true
 	}
 	f.mglv4.Range(func(_ string, pair kvpair) bool {
-		return cbWrap(*pair.a.(*lbmap.MaglevOuterKey), pair.b.(lbmap.MaglevOuterVal), false)
+		return cbWrap(*pair.a.(*MaglevOuterKey), pair.b.(MaglevOuterVal), false)
 	})
 	f.mglv6.Range(func(_ string, pair kvpair) bool {
-		return cbWrap(*pair.a.(*lbmap.MaglevOuterKey), pair.b.(lbmap.MaglevOuterVal), true)
+		return cbWrap(*pair.a.(*MaglevOuterKey), pair.b.(MaglevOuterVal), true)
 	})
 	return err
 }
@@ -1093,10 +1141,10 @@ func (f *FakeLBMaps) DumpMaglev(cb func(lbmap.MaglevOuterKey, lbmap.MaglevOuterV
 func (f *FakeLBMaps) DeleteSockRevNat(cookie uint64, addr net.IP, port uint16) error {
 	var key bpf.MapKey
 	if addr.To4() != nil {
-		key4 := lbmap.NewSockRevNat4Key(cookie, addr, port)
+		key4 := NewSockRevNat4Key(cookie, addr, port)
 		key = key4
 	} else {
-		key6 := lbmap.NewSockRevNat6Key(cookie, addr, port)
+		key6 := NewSockRevNat6Key(cookie, addr, port)
 		key = key6
 	}
 	return f.sockRevNat.delete(key)
@@ -1107,17 +1155,17 @@ func (f *FakeLBMaps) UpdateSockRevNat(cookie uint64, addr net.IP, port uint16, r
 	var key bpf.MapKey
 	var value bpf.MapValue
 	if addr.To4() != nil {
-		key4 := lbmap.NewSockRevNat4Key(cookie, addr, port)
+		key4 := NewSockRevNat4Key(cookie, addr, port)
 		key = key4
-		value = &lbmap.SockRevNat4Value{
+		value = &SockRevNat4Value{
 			Address:     key4.Address,
 			Port:        key4.Port,
 			RevNatIndex: revNatIndex,
 		}
 	} else {
-		key6 := lbmap.NewSockRevNat6Key(cookie, addr, port)
+		key6 := NewSockRevNat6Key(cookie, addr, port)
 		key = key6
-		value = &lbmap.SockRevNat6Value{
+		value = &SockRevNat6Value{
 			Address:     key6.Address,
 			Port:        key6.Port,
 			RevNatIndex: revNatIndex,
@@ -1130,10 +1178,10 @@ func (f *FakeLBMaps) UpdateSockRevNat(cookie uint64, addr net.IP, port uint16, r
 func (f *FakeLBMaps) ExistsSockRevNat(cookie uint64, addr net.IP, port uint16) bool {
 	var key bpf.MapKey
 	if addr.To4() != nil {
-		key4 := lbmap.NewSockRevNat4Key(cookie, addr, port)
+		key4 := NewSockRevNat4Key(cookie, addr, port)
 		key = key4
 	} else {
-		key6 := lbmap.NewSockRevNat6Key(cookie, addr, port)
+		key6 := NewSockRevNat6Key(cookie, addr, port)
 		key = key6
 	}
 	return f.sockRevNat.exists(key)
@@ -1170,35 +1218,35 @@ func (s *mapSnapshots) snapshot(lbmaps LBMaps) error {
 	s.mu.Lock()
 	defer s.mu.Unlock()
 
-	svcCB := func(svcKey lbmap.ServiceKey, svcValue lbmap.ServiceValue) {
+	svcCB := func(svcKey ServiceKey, svcValue ServiceValue) {
 		s.services = append(s.services, mapKeyValue{svcKey, svcValue})
 	}
 	if err := lbmaps.DumpService(svcCB); err != nil {
 		return fmt.Errorf("DumpService: %w", err)
 	}
 
-	beCB := func(beKey lbmap.BackendKey, beValue lbmap.BackendValue) {
+	beCB := func(beKey BackendKey, beValue BackendValue) {
 		s.backends = append(s.backends, mapKeyValue{beKey, beValue})
 	}
 	if err := lbmaps.DumpBackend(beCB); err != nil {
 		return fmt.Errorf("DumpBackend: %w", err)
 	}
 
-	revCB := func(revKey lbmap.RevNatKey, revValue lbmap.RevNatValue) {
+	revCB := func(revKey RevNatKey, revValue RevNatValue) {
 		s.revNat = append(s.revNat, mapKeyValue{revKey, revValue})
 	}
 	if err := lbmaps.DumpRevNat(revCB); err != nil {
 		return fmt.Errorf("DumpRevNat: %w", err)
 	}
 
-	affCB := func(affKey *lbmap.AffinityMatchKey, affValue *lbmap.AffinityMatchValue) {
+	affCB := func(affKey *AffinityMatchKey, affValue *AffinityMatchValue) {
 		s.affinity = append(s.revNat, mapKeyValue{affKey, affValue})
 	}
 	if err := lbmaps.DumpAffinityMatch(affCB); err != nil {
 		return fmt.Errorf("DumpAffinityMatch: %w", err)
 	}
 
-	srcRangeCB := func(key lbmap.SourceRangeKey, value *lbmap.SourceRangeValue) {
+	srcRangeCB := func(key SourceRangeKey, value *SourceRangeValue) {
 		s.srcRange = append(s.srcRange, mapKeyValue{key, value})
 	}
 	if err := lbmaps.DumpSourceRange(srcRangeCB); err != nil {
@@ -1212,19 +1260,19 @@ func (s *mapSnapshots) restore(lbmaps LBMaps) (err error) {
 	defer s.mu.Unlock()
 
 	for _, kv := range s.services {
-		err = errors.Join(err, lbmaps.UpdateService(kv.key.(lbmap.ServiceKey), kv.value.(lbmap.ServiceValue)))
+		err = errors.Join(err, lbmaps.UpdateService(kv.key.(ServiceKey), kv.value.(ServiceValue)))
 	}
 	for _, kv := range s.backends {
-		err = errors.Join(err, lbmaps.UpdateBackend(kv.key.(lbmap.BackendKey), kv.value.(lbmap.BackendValue)))
+		err = errors.Join(err, lbmaps.UpdateBackend(kv.key.(BackendKey), kv.value.(BackendValue)))
 	}
 	for _, kv := range s.revNat {
-		err = errors.Join(err, lbmaps.UpdateRevNat(kv.key.(lbmap.RevNatKey), kv.value.(lbmap.RevNatValue)))
+		err = errors.Join(err, lbmaps.UpdateRevNat(kv.key.(RevNatKey), kv.value.(RevNatValue)))
 	}
 	for _, kv := range s.affinity {
-		err = errors.Join(err, lbmaps.UpdateAffinityMatch(kv.key.(*lbmap.AffinityMatchKey), kv.value.(*lbmap.AffinityMatchValue)))
+		err = errors.Join(err, lbmaps.UpdateAffinityMatch(kv.key.(*AffinityMatchKey), kv.value.(*AffinityMatchValue)))
 	}
 	for _, kv := range s.srcRange {
-		err = errors.Join(err, lbmaps.UpdateSourceRange(kv.key.(lbmap.SourceRangeKey), kv.value.(*lbmap.SourceRangeValue)))
+		err = errors.Join(err, lbmaps.UpdateSourceRange(kv.key.(SourceRangeKey), kv.value.(*SourceRangeValue)))
 	}
 	return
 }
