+++ pkg/loadbalancer/maps/test_utils.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/cilium/cilium/pkg/byteorder"
 	"github.com/cilium/cilium/pkg/loadbalancer"
-	"github.com/cilium/cilium/pkg/maps/lbmap"
 )
 
 // MapDump is a dump of a BPF map. These are generated by the dump() method, which
@@ -51,7 +50,7 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 	}
 
 	// BackendID corresponds to a union type in bpf/lib/common.h, and we stringify it differently depending on the case of the union.
-	stringFromBackendID := func(svcValue lbmap.ServiceValue, slot int) string {
+	stringFromBackendID := func(svcValue ServiceValue, slot int) string {
 		if slot != 0 {
 			return fmt.Sprintf("BEID=%s", sanitizeID(svcValue.GetBackendID(), sanitizeIDs))
 		}
@@ -64,7 +63,7 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		)
 	}
 
-	svcCB := func(svcKey lbmap.ServiceKey, svcValue lbmap.ServiceValue) {
+	svcCB := func(svcKey ServiceKey, svcValue ServiceValue) {
 		svcKey = svcKey.ToHost()
 		svcValue = svcValue.ToHost()
 		addr := svcKey.GetAddress()
@@ -89,7 +88,7 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		panic(err)
 	}
 
-	beCB := func(beKey lbmap.BackendKey, beValue lbmap.BackendValue) {
+	beCB := func(beKey BackendKey, beValue BackendValue) {
 		beValue = beValue.ToHost()
 		addr := beValue.GetAddress()
 		addrS := customizeAddr(addr, beValue.GetPort())
@@ -105,17 +104,17 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		panic(err)
 	}
 
-	revCB := func(revKey lbmap.RevNatKey, revValue lbmap.RevNatValue) {
+	revCB := func(revKey RevNatKey, revValue RevNatValue) {
 		revKey = revKey.ToHost()
 		revValue = revValue.ToHost()
 
 		var addr string
 
 		switch v := revValue.(type) {
-		case *lbmap.RevNat4Value:
+		case *RevNat4Value:
 			addr = customizeAddr(v.Address.IP(), v.Port)
 
-		case *lbmap.RevNat6Value:
+		case *RevNat6Value:
 			addr = customizeAddr(v.Address.IP(), v.Port)
 		}
 
@@ -128,7 +127,7 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		panic(err)
 	}
 
-	affCB := func(affKey *lbmap.AffinityMatchKey, _ *lbmap.AffinityMatchValue) {
+	affCB := func(affKey *AffinityMatchKey, _ *AffinityMatchValue) {
 		affKey = affKey.ToHost()
 		out = append(out, fmt.Sprintf("AFF: ID=%s BEID=%d",
 			sanitizeID(affKey.RevNATID, sanitizeIDs),
@@ -140,7 +139,7 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		panic(err)
 	}
 
-	srcRangeCB := func(key lbmap.SourceRangeKey, _ *lbmap.SourceRangeValue) {
+	srcRangeCB := func(key SourceRangeKey, _ *SourceRangeValue) {
 		key = key.ToHost()
 		out = append(out, fmt.Sprintf("SRCRANGE: ID=%s CIDR=%s",
 			sanitizeID(key.GetRevNATID(), sanitizeIDs),
@@ -151,8 +150,8 @@ func DumpLBMaps(lbmaps LBMaps, sanitizeIDs bool, customizeAddr func(net.IP, uint
 		panic(err)
 	}
 
-	maglevCB := func(key lbmap.MaglevOuterKey, _ lbmap.MaglevOuterVal, _ lbmap.MaglevInnerKey, innerValue *lbmap.MaglevInnerVal, _ bool) {
-		key = lbmap.MaglevOuterKey{
+	maglevCB := func(key MaglevOuterKey, _ MaglevOuterVal, _ MaglevInnerKey, innerValue *MaglevInnerVal, _ bool) {
+		key = MaglevOuterKey{
 			RevNatID: byteorder.NetworkToHost16(key.RevNatID),
 		}
 		idCounts := make(map[loadbalancer.BackendID]int)
