+++ pkg/loadbalancer/reconciler/bpf_reconciler.go
@@ -29,7 +29,6 @@ import (
 	"github.com/cilium/cilium/pkg/loadbalancer/writer"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/maglev"
-	"github.com/cilium/cilium/pkg/maps/lbmap"
 	"github.com/cilium/cilium/pkg/time"
 	"github.com/cilium/cilium/pkg/u8proto"
 )
@@ -213,7 +212,7 @@ func (ops *BPFOps) ResetAndRestore() (err error) {
 	ops.prevSourceRanges = map[loadbalancer.L3n4Addr]sets.Set[netip.Prefix]{}
 
 	// Restore backend IDs
-	err = ops.LBMaps.DumpBackend(func(key lbmap.BackendKey, value lbmap.BackendValue) {
+	err = ops.LBMaps.DumpBackend(func(key maps.BackendKey, value maps.BackendValue) {
 		value = value.ToHost()
 		ops.backendIDAlloc.addID(beValueToAddr(value), loadbalancer.ID(key.GetID()))
 		ops.restoredBackendIDs.Insert(key.GetID())
@@ -223,8 +222,8 @@ func (ops *BPFOps) ResetAndRestore() (err error) {
 	}
 
 	// Gather all services key'd by address.
-	serviceSlots := map[loadbalancer.L3n4Addr][]lbmap.ServiceValue{}
-	err = ops.LBMaps.DumpService(func(key lbmap.ServiceKey, value lbmap.ServiceValue) {
+	serviceSlots := map[loadbalancer.L3n4Addr][]maps.ServiceValue{}
+	err = ops.LBMaps.DumpService(func(key maps.ServiceKey, value maps.ServiceValue) {
 		key = key.ToHost()
 		value = value.ToHost()
 		addr := svcKeyToAddr(key)
@@ -268,7 +267,7 @@ func (ops *BPFOps) ResetAndRestore() (err error) {
 	return nil
 }
 
-func svcKeyToAddr(svcKey lbmap.ServiceKey) loadbalancer.L3n4Addr {
+func svcKeyToAddr(svcKey maps.ServiceKey) loadbalancer.L3n4Addr {
 	feIP := svcKey.GetAddress()
 	feAddrCluster := cmtypes.MustAddrClusterFromIP(feIP)
 	proto := loadbalancer.NewL4TypeFromNumber(svcKey.GetProtocol())
@@ -276,7 +275,7 @@ func svcKeyToAddr(svcKey lbmap.ServiceKey) loadbalancer.L3n4Addr {
 	return *feL3n4Addr
 }
 
-func beValueToAddr(beValue lbmap.BackendValue) loadbalancer.L3n4Addr {
+func beValueToAddr(beValue maps.BackendValue) loadbalancer.L3n4Addr {
 	beIP := beValue.GetAddress()
 	beAddrCluster := cmtypes.MustAddrClusterFromIP(beIP)
 	proto := loadbalancer.NewL4TypeFromNumber(beValue.GetProtocol())
@@ -357,7 +356,7 @@ func (ops *BPFOps) deleteFrontend(fe *loadbalancer.Frontend) error {
 
 	// Delete Maglev.
 	if ops.useMaglev(fe) {
-		if err := ops.LBMaps.DeleteMaglev(lbmap.MaglevOuterKey{RevNatID: uint16(feID)}, fe.Address.IsIPv6()); err != nil {
+		if err := ops.LBMaps.DeleteMaglev(maps.MaglevOuterKey{RevNatID: uint16(feID)}, fe.Address.IsIPv6()); err != nil {
 			return fmt.Errorf("ops.LBMaps.DeleteMaglev failed: %w", err)
 		}
 	}
@@ -380,8 +379,8 @@ func (ops *BPFOps) deleteFrontend(fe *loadbalancer.Frontend) error {
 		ops.releaseBackend(orphanState.id, orphanState.addr)
 	}
 
-	var svcKey lbmap.ServiceKey
-	var revNatKey lbmap.RevNatKey
+	var svcKey maps.ServiceKey
+	var revNatKey maps.RevNatKey
 
 	ip := fe.Address.AddrCluster.AsNetIP()
 	proto, err := u8proto.ParseProtocol(fe.Address.Protocol)
@@ -389,11 +388,11 @@ func (ops *BPFOps) deleteFrontend(fe *loadbalancer.Frontend) error {
 		return fmt.Errorf("invalid L4 protocol %q: %w", fe.Address.Protocol, err)
 	}
 	if fe.Address.IsIPv6() {
-		svcKey = lbmap.NewService6Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
-		revNatKey = lbmap.NewRevNat6Key(uint16(feID))
+		svcKey = maps.NewService6Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
+		revNatKey = maps.NewRevNat6Key(uint16(feID))
 	} else {
-		svcKey = lbmap.NewService4Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
-		revNatKey = lbmap.NewRevNat4Key(uint16(feID))
+		svcKey = maps.NewService4Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
+		revNatKey = maps.NewRevNat4Key(uint16(feID))
 	}
 
 	// Delete all slots including master.
@@ -438,8 +437,8 @@ func (ops *BPFOps) deleteFrontend(fe *loadbalancer.Frontend) error {
 }
 
 func (ops *BPFOps) pruneServiceMaps() error {
-	toDelete := []lbmap.ServiceKey{}
-	svcCB := func(svcKey lbmap.ServiceKey, svcValue lbmap.ServiceValue) {
+	toDelete := []maps.ServiceKey{}
+	svcCB := func(svcKey maps.ServiceKey, svcValue maps.ServiceValue) {
 		svcKey = svcKey.ToHost()
 		svcValue = svcValue.ToHost()
 		ac, ok := cmtypes.AddrClusterFromIP(svcKey.GetAddress())
@@ -484,8 +483,8 @@ func (ops *BPFOps) pruneServiceMaps() error {
 }
 
 func (ops *BPFOps) pruneBackendMaps() error {
-	toDelete := []lbmap.BackendKey{}
-	beCB := func(beKey lbmap.BackendKey, beValue lbmap.BackendValue) {
+	toDelete := []maps.BackendKey{}
+	beCB := func(beKey maps.BackendKey, beValue maps.BackendValue) {
 		beValue = beValue.ToHost()
 		addr := beValueToAddr(beValue)
 		if _, ok := ops.backendStates[addr]; !ok {
@@ -533,8 +532,8 @@ func (ops *BPFOps) pruneRestoredIDs() error {
 }
 
 func (ops *BPFOps) pruneRevNat() error {
-	toDelete := []lbmap.RevNatKey{}
-	cb := func(key lbmap.RevNatKey, value lbmap.RevNatValue) {
+	toDelete := []maps.RevNatKey{}
+	cb := func(key maps.RevNatKey, value maps.RevNatValue) {
 		key = key.ToHost()
 		if _, ok := ops.serviceIDAlloc.entitiesID[loadbalancer.ID(key.GetKey())]; !ok {
 			ops.log.Debug("pruneRevNat: enqueing for deletion", logfields.ID, key.GetKey())
@@ -555,8 +554,8 @@ func (ops *BPFOps) pruneRevNat() error {
 }
 
 func (ops *BPFOps) pruneSourceRanges() error {
-	toDelete := []lbmap.SourceRangeKey{}
-	cb := func(key lbmap.SourceRangeKey, value *lbmap.SourceRangeValue) {
+	toDelete := []maps.SourceRangeKey{}
+	cb := func(key maps.SourceRangeKey, value *maps.SourceRangeValue) {
 		key = key.ToHost()
 
 		// A SourceRange is OK if there's a service with this ID and the
@@ -593,11 +592,11 @@ func (ops *BPFOps) pruneSourceRanges() error {
 
 func (ops *BPFOps) pruneMaglev() error {
 	type outerKeyWithIPVersion struct {
-		lbmap.MaglevOuterKey
+		maps.MaglevOuterKey
 		ipv6 bool
 	}
 	toDelete := []outerKeyWithIPVersion{}
-	cb := func(key lbmap.MaglevOuterKey, _ lbmap.MaglevOuterVal, _ lbmap.MaglevInnerKey, _ *lbmap.MaglevInnerVal, ipv6 bool) {
+	cb := func(key maps.MaglevOuterKey, _ maps.MaglevOuterVal, _ maps.MaglevInnerKey, _ *maps.MaglevInnerVal, ipv6 bool) {
 		if _, ok := ops.serviceIDAlloc.entitiesID[loadbalancer.ID(key.RevNatID)]; !ok {
 			ops.log.Debug("pruneMaglev: enqueing for deletion", logfields.ID, key.RevNatID)
 			toDelete = append(toDelete, outerKeyWithIPVersion{key, ipv6})
@@ -714,8 +713,8 @@ func (ops *BPFOps) updateFrontend(fe *loadbalancer.Frontend) error {
 	}
 	fe.ID = loadbalancer.ServiceID(feID)
 
-	var svcKey lbmap.ServiceKey
-	var svcVal lbmap.ServiceValue
+	var svcKey maps.ServiceKey
+	var svcVal maps.ServiceValue
 
 	proto, err := u8proto.ParseProtocol(fe.Address.Protocol)
 	if err != nil {
@@ -724,11 +723,11 @@ func (ops *BPFOps) updateFrontend(fe *loadbalancer.Frontend) error {
 
 	ip := fe.Address.AddrCluster.AsNetIP()
 	if fe.Address.IsIPv6() {
-		svcKey = lbmap.NewService6Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
-		svcVal = &lbmap.Service6Value{}
+		svcKey = maps.NewService6Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
+		svcVal = &maps.Service6Value{}
 	} else {
-		svcKey = lbmap.NewService4Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
-		svcVal = &lbmap.Service4Value{}
+		svcKey = maps.NewService4Key(ip, fe.Address.Port, proto, fe.Address.Scope, 0)
+		svcVal = &maps.Service4Value{}
 	}
 
 	svcType := fe.Type
@@ -895,7 +894,7 @@ func (ops *BPFOps) updateFrontend(fe *loadbalancer.Frontend) error {
 		ops.prevSourceRanges[fe.Address] = prevSourceRanges
 	}
 	orphanSourceRanges := prevSourceRanges.Clone()
-	srcRangeValue := &lbmap.SourceRangeValue{}
+	srcRangeValue := &maps.SourceRangeValue{}
 	for _, prefix := range fe.Service.SourceRanges {
 		if prefix.Addr().Is6() != fe.Address.IsIPv6() {
 			continue
@@ -1008,7 +1007,7 @@ func (ops *BPFOps) useMaglev(fe *loadbalancer.Frontend) bool {
 	}
 }
 
-func (ops *BPFOps) upsertService(svcKey lbmap.ServiceKey, svcVal lbmap.ServiceValue) error {
+func (ops *BPFOps) upsertService(svcKey maps.ServiceKey, svcVal maps.ServiceValue) error {
 	var err error
 	svcKey = svcKey.ToNetwork()
 	svcVal = svcVal.ToNetwork()
@@ -1024,7 +1023,7 @@ func (ops *BPFOps) upsertService(svcKey lbmap.ServiceKey, svcVal lbmap.ServiceVa
 	return err
 }
 
-func (ops *BPFOps) upsertMaster(svcKey lbmap.ServiceKey, svcVal lbmap.ServiceValue, fe *loadbalancer.Frontend, activeBackends, inactiveBackends int) error {
+func (ops *BPFOps) upsertMaster(svcKey maps.ServiceKey, svcVal maps.ServiceValue, fe *loadbalancer.Frontend, activeBackends, inactiveBackends int) error {
 	svcVal.SetCount(activeBackends)
 	svcVal.SetQCount(inactiveBackends)
 	svcKey.SetBackendSlot(0)
@@ -1045,7 +1044,7 @@ func (ops *BPFOps) upsertMaster(svcKey lbmap.ServiceKey, svcVal lbmap.ServiceVal
 	return ops.upsertService(svcKey, svcVal)
 }
 
-func (ops *BPFOps) cleanupSlots(svcKey lbmap.ServiceKey, oldCount, newCount int) error {
+func (ops *BPFOps) cleanupSlots(svcKey maps.ServiceKey, oldCount, newCount int) error {
 	for i := newCount; i < oldCount; i++ {
 		svcKey.SetBackendSlot(i + 1)
 		err := ops.LBMaps.DeleteService(svcKey.ToNetwork())
@@ -1057,20 +1056,20 @@ func (ops *BPFOps) cleanupSlots(svcKey lbmap.ServiceKey, oldCount, newCount int)
 }
 
 func (ops *BPFOps) upsertBackend(id loadbalancer.BackendID, be *loadbalancer.BackendParams) (err error) {
-	var lbbe lbmap.Backend
+	var lbbe maps.Backend
 	proto, err := u8proto.ParseProtocol(be.Address.Protocol)
 	if err != nil {
 		return fmt.Errorf("invalid L4 protocol %q: %w", be.Address.Protocol, err)
 	}
 
 	if be.Address.AddrCluster.Is6() {
-		lbbe, err = lbmap.NewBackend6V3(id, be.Address.AddrCluster, be.Address.Port, proto,
+		lbbe, err = maps.NewBackend6V3(id, be.Address.AddrCluster, be.Address.Port, proto,
 			be.State, ops.extCfg.GetZoneID(be.Zone))
 		if err != nil {
 			return err
 		}
 	} else {
-		lbbe, err = lbmap.NewBackend4V3(id, be.Address.AddrCluster, be.Address.Port, proto,
+		lbbe, err = maps.NewBackend4V3(id, be.Address.AddrCluster, be.Address.Port, proto,
 			be.State, ops.extCfg.GetZoneID(be.Zone))
 		if err != nil {
 			return err
@@ -1083,11 +1082,11 @@ func (ops *BPFOps) upsertBackend(id loadbalancer.BackendID, be *loadbalancer.Bac
 }
 
 func (ops *BPFOps) deleteBackend(ipv6 bool, id loadbalancer.BackendID) error {
-	var key lbmap.BackendKey
+	var key maps.BackendKey
 	if ipv6 {
-		key = lbmap.NewBackend6KeyV3(id)
+		key = maps.NewBackend6KeyV3(id)
 	} else {
-		key = lbmap.NewBackend4KeyV3(id)
+		key = maps.NewBackend4KeyV3(id)
 	}
 	err := ops.LBMaps.DeleteBackend(key)
 	if err != nil {
@@ -1097,24 +1096,24 @@ func (ops *BPFOps) deleteBackend(ipv6 bool, id loadbalancer.BackendID) error {
 }
 
 func (ops *BPFOps) upsertAffinityMatch(id loadbalancer.ID, beID loadbalancer.BackendID) error {
-	key := &lbmap.AffinityMatchKey{
+	key := &maps.AffinityMatchKey{
 		BackendID: beID,
 		RevNATID:  uint16(id),
 	}
-	var value lbmap.AffinityMatchValue
+	var value maps.AffinityMatchValue
 	return ops.LBMaps.UpdateAffinityMatch(key.ToNetwork(), &value)
 }
 
 func (ops *BPFOps) deleteAffinityMatch(id loadbalancer.ID, beID loadbalancer.BackendID) error {
-	key := &lbmap.AffinityMatchKey{
+	key := &maps.AffinityMatchKey{
 		BackendID: beID,
 		RevNATID:  uint16(id),
 	}
 	return ops.LBMaps.DeleteAffinityMatch(key.ToNetwork())
 }
 
-func (ops *BPFOps) upsertRevNat(id loadbalancer.ID, svcKey lbmap.ServiceKey, svcVal lbmap.ServiceValue) error {
-	zeroValue := svcVal.New().(lbmap.ServiceValue)
+func (ops *BPFOps) upsertRevNat(id loadbalancer.ID, svcKey maps.ServiceKey, svcVal maps.ServiceValue) error {
+	zeroValue := svcVal.New().(maps.ServiceValue)
 	zeroValue.SetRevNat(int(id))
 	revNATKey := zeroValue.RevNatKey()
 	revNATValue := svcKey.RevNatValue()
@@ -1137,7 +1136,7 @@ type backendWithRevision struct {
 
 func (ops *BPFOps) updateMaglev(fe *loadbalancer.Frontend, feID loadbalancer.ID, activeBackends []backendWithRevision) error {
 	if len(activeBackends) == 0 {
-		if err := ops.LBMaps.DeleteMaglev(lbmap.MaglevOuterKey{RevNatID: uint16(feID)}, fe.Address.IsIPv6()); err != nil {
+		if err := ops.LBMaps.DeleteMaglev(maps.MaglevOuterKey{RevNatID: uint16(feID)}, fe.Address.IsIPv6()); err != nil {
 			return fmt.Errorf("ops.LBMaps.DeleteMaglev failed: %w", err)
 		}
 		return nil
@@ -1146,7 +1145,7 @@ func (ops *BPFOps) updateMaglev(fe *loadbalancer.Frontend, feID loadbalancer.ID,
 	if err != nil {
 		return fmt.Errorf("ops.computeMaglevTable failed: %w", err)
 	}
-	if err := ops.LBMaps.UpdateMaglev(lbmap.MaglevOuterKey{RevNatID: uint16(feID)}, maglevTable, fe.Address.IsIPv6()); err != nil {
+	if err := ops.LBMaps.UpdateMaglev(maps.MaglevOuterKey{RevNatID: uint16(feID)}, maglevTable, fe.Address.IsIPv6()); err != nil {
 		return fmt.Errorf("ops.LBMaps.UpdateMaglev failed: %w", err)
 	}
 	return nil
@@ -1433,7 +1432,7 @@ func newID(svc loadbalancer.L3n4Addr, id loadbalancer.ID) *loadbalancer.L3n4Addr
 	}
 }
 
-func srcRangeKey(cidr netip.Prefix, revNATID uint16, ipv6 bool) lbmap.SourceRangeKey {
+func srcRangeKey(cidr netip.Prefix, revNATID uint16, ipv6 bool) maps.SourceRangeKey {
 	const (
 		lpmPrefixLen4 = 16 + 16 // sizeof(SourceRangeKey4.RevNATID)+sizeof(SourceRangeKey4.Pad)
 		lpmPrefixLen6 = 16 + 16 // sizeof(SourceRangeKey6.RevNATID)+sizeof(SourceRangeKey6.Pad)
@@ -1441,12 +1440,12 @@ func srcRangeKey(cidr netip.Prefix, revNATID uint16, ipv6 bool) lbmap.SourceRang
 	ones := cidr.Bits()
 	id := byteorder.HostToNetwork16(revNATID)
 	if ipv6 {
-		key := &lbmap.SourceRangeKey6{PrefixLen: uint32(ones) + lpmPrefixLen6, RevNATID: id}
+		key := &maps.SourceRangeKey6{PrefixLen: uint32(ones) + lpmPrefixLen6, RevNATID: id}
 		as16 := cidr.Addr().As16()
 		copy(key.Address[:], as16[:])
 		return key
 	} else {
-		key := &lbmap.SourceRangeKey4{PrefixLen: uint32(ones) + lpmPrefixLen4, RevNATID: id}
+		key := &maps.SourceRangeKey4{PrefixLen: uint32(ones) + lpmPrefixLen4, RevNATID: id}
 		as4 := cidr.Addr().As4()
 		copy(key.Address[:], as4[:])
 		return key
