+++ pkg/maps/lbmap/maglev_outer_map.go
@@ -9,9 +9,14 @@ import (
 	"log/slog"
 	"unsafe"
 
-	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/byteorder"
 	"github.com/cilium/cilium/pkg/ebpf"
+	"github.com/cilium/cilium/pkg/loadbalancer/maps"
+)
+
+type (
+	MaglevOuterKey = maps.MaglevOuterKey
+	MaglevOuterVal = maps.MaglevOuterVal
 )
 
 // MaglevOuterMap represents a Maglev outer map.
@@ -23,37 +28,10 @@ type MaglevOuterMap struct {
 // UpdateService sets the given inner map to be the Maglev lookup table for
 // the service with the given id.
 func (m *MaglevOuterMap) UpdateService(id uint16, inner *MaglevInnerMap) error {
-	key := MaglevOuterKey{RevNatID: id}.toNetwork()
+	key := MaglevOuterKey{RevNatID: id}.ToNetwork()
 	val := MaglevOuterVal{FD: uint32(inner.FD())}
 	return m.Map.Update(key, val, 0)
 }
-
-// MaglevOuterKey is the key of a maglev outer map.
-type MaglevOuterKey struct {
-	RevNatID uint16
-}
-
-// New and String implement bpf.MapKey
-func (k *MaglevOuterKey) New() bpf.MapKey { return &MaglevOuterKey{} }
-func (k *MaglevOuterKey) String() string  { return fmt.Sprintf("%d", k.RevNatID) }
-
-var _ bpf.MapKey = &MaglevOuterKey{}
-
-// toNetwork converts a maglev outer map's key to network byte order.
-// The key is in network byte order in the eBPF maps.
-func (k MaglevOuterKey) toNetwork() MaglevOuterKey {
-	return MaglevOuterKey{
-		RevNatID: byteorder.HostToNetwork16(k.RevNatID),
-	}
-}
-
-// MaglevOuterVal is the value of a maglev outer map.
-type MaglevOuterVal struct {
-	FD uint32
-}
-
-func (v *MaglevOuterVal) New() bpf.MapValue { return &MaglevOuterVal{} }
-func (k *MaglevOuterVal) String() string    { return fmt.Sprintf("%d", k.FD) }
 
 // NewMaglevOuterMap returns a new object representing a maglev outer map.
 func NewMaglevOuterMap(logger *slog.Logger, name string, maxEntries int, tableSize uint32, innerMap *ebpf.MapSpec) (*MaglevOuterMap, error) {
@@ -114,7 +92,7 @@ func (m *MaglevOuterMap) TableSize() (uint32, error) {
 
 // GetService gets the maglev backend lookup table for the given service id.
 func (m *MaglevOuterMap) GetService(id uint16) (*MaglevInnerMap, error) {
-	key := MaglevOuterKey{RevNatID: id}.toNetwork()
+	key := MaglevOuterKey{RevNatID: id}.ToNetwork()
 	var val MaglevOuterVal
 
 	err := m.Lookup(key, &val)
